refactor(kwin-effect): extract drag/snap helpers to plasmazoneseffect/drag_snap.cpp
Pulls 10 drag/snap methods out of plasmazoneseffect.cpp: borderActivated,
callResolveWindowRestore, callEndDrag, callCancelSnap, tryAsyncSnapCall,
repaintSnapRegions, applySnapGeometry, slotRestoreSizeDuringDrag,
slotSnapAssistReady, slotDragPolicyChanged. Also moves the
EndDragTimeoutMs anon-namespace constant. Drops the main file from 2452
to 1844 lines.
