+++ kwin-effect/plasmazoneseffect/drag_snap.cpp
@@ -1,0 +1,643 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#include "../plasmazoneseffect.h"
+
+#include <PhosphorAnimation/ProfilePaths.h>
+#include <PhosphorProtocol/ClientHelpers.h>
+#include <PhosphorProtocol/ServiceConstants.h>
+#include <PhosphorProtocol/WireTypes.h>
+
+#include <effect/effecthandler.h>
+#include <window.h>
+
+#include <QDBusConnection>
+#include <QDBusMessage>
+#include <QDBusPendingCall>
+#include <QDBusPendingCallWatcher>
+#include <QDBusPendingReply>
+#include <QLoggingCategory>
+#include <QPointer>
+#include <QTimer>
+
+#include <memory>
+
+#include "../autotilehandler.h"
+#include "../dragtracker.h"
+#include "../snapassisthandler.h"
+#include "../windowanimator.h"
+
+namespace PlasmaZones {
+
+Q_DECLARE_LOGGING_CATEGORY(lcEffect)
+
+namespace {
+// Upper bound on how long the effect waits for the daemon's endDrag reply.
+// If the daemon is blocked (layout recompute, overlay teardown, heavy
+// handler), exceeding this budget means the compositor would otherwise
+// stall waiting on a reply that may never come. On expiry the window is
+// left at its release position and a warning is logged.
+constexpr int EndDragTimeoutMs = 500;
+} // namespace
+
+bool PlasmaZonesEffect::borderActivated(KWin::ElectricBorder border)
+{
+    Q_UNUSED(border)
+    // We no longer reserve edges, so this callback won't be triggered by our effect.
+    // The daemon handles disabling Quick Tile via KWin config.
+    return false;
+}
+
+void PlasmaZonesEffect::callResolveWindowRestore(KWin::EffectWindow* window, std::function<void()> onComplete)
+{
+    if (!window) {
+        if (onComplete)
+            onComplete();
+        return;
+    }
+
+    if (!isDaemonReady("resolve window restore")) {
+        if (onComplete)
+            onComplete();
+        return;
+    }
+
+    QString windowId = getWindowId(window);
+    QString screenId = getWindowScreenId(window);
+    bool sticky = isWindowSticky(window);
+
+    QPointer<KWin::EffectWindow> safeWindow = window;
+
+    // Single D-Bus call — daemon runs the full appRule → persisted → emptyZone → lastZone chain
+    // skipAnimation=true: window is being restored to its snap position on startup/reopen,
+    // so teleport directly instead of sliding from KWin's saved position.
+    // storePreSnap=false: the window is already at its snap/zone position (from before
+    // daemon restart or from KWin session restore), so its current frameGeometry is the
+    // zone geometry — NOT the free-floating geometry. Storing it as pre-tile would cause
+    // float toggle to restore to the zone geometry instead of the original free-floating position.
+    tryAsyncSnapCall(PhosphorProtocol::Service::Interface::Snap, QStringLiteral("resolveWindowRestore"),
+                     {windowId, screenId, sticky}, safeWindow, windowId, false, nullptr, nullptr,
+                     /*skipAnimation=*/true, onComplete);
+}
+
+// The kwin-effect no longer calls the legacy dragStarted D-Bus method;
+// beginDrag sets up snap-path state internally on the daemon side, so
+// there's only one code path into the drag state machine.
+
+// The dragMoved lambda sends updateDragCursor directly via
+// ClientHelpers::fireAndForget. Single entry point for hot-path cursor updates.
+
+void PlasmaZonesEffect::callEndDrag(KWin::EffectWindow* window, const QString& windowId, bool cancelled)
+{
+    // Single entry point for drag-end dispatch.
+    // Sends endDrag, receives a PhosphorProtocol::DragOutcome, and applies exactly the
+    // action the daemon decided. Replaces callDragStopped (whose reply
+    // shape was a 9-tuple of out-params) with a typed struct.
+    QPointF cursorAtRelease = m_dragTracker->lastCursorPos();
+
+    QDBusMessage msg =
+        QDBusMessage::createMethodCall(PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                                       PhosphorProtocol::Service::Interface::WindowDrag, QStringLiteral("endDrag"));
+    msg << windowId << static_cast<int>(cursorAtRelease.x()) << static_cast<int>(cursorAtRelease.y())
+        << static_cast<int>(m_currentModifiers) << static_cast<int>(m_currentMouseButtons) << cancelled;
+    QDBusPendingCall pendingCall = QDBusConnection::sessionBus().asyncCall(msg);
+
+    QPointer<KWin::EffectWindow> safeWindow = window;
+    auto* watcher = new QDBusPendingCallWatcher(pendingCall, this);
+
+    // Pair the watcher with a timeout. If the daemon is blocked (layout
+    // recompute, overlay teardown, heavy handler), the compositor would
+    // otherwise wait indefinitely for a reply that may never come. The
+    // shared `handled` flag guarantees exactly-once handling: whichever
+    // fires first (reply or timeout) takes the transition, the other path
+    // is a no-op. Deleting the watcher does NOT cancel the underlying
+    // QDBusPendingCall — any late reply is silently discarded by Qt.
+    auto handled = std::make_shared<bool>(false);
+    QTimer* timeoutTimer = new QTimer(this);
+    timeoutTimer->setSingleShot(true);
+    connect(timeoutTimer, &QTimer::timeout, this, [this, windowId, handled, watcher, timeoutTimer]() {
+        if (*handled) {
+            return;
+        }
+        *handled = true;
+        qCWarning(lcEffect) << "endDrag timed out after" << EndDragTimeoutMs
+                            << "ms; daemon unresponsive. Leaving window" << windowId << "at release position.";
+        watcher->deleteLater();
+        timeoutTimer->deleteLater();
+    });
+    timeoutTimer->start(EndDragTimeoutMs);
+
+    connect(watcher, &QDBusPendingCallWatcher::finished, this,
+            [this, safeWindow, windowId, handled, timeoutTimer](QDBusPendingCallWatcher* w) {
+                w->deleteLater();
+                if (*handled) {
+                    // Timeout already fired; this is a late reply — discard.
+                    return;
+                }
+                *handled = true;
+                timeoutTimer->stop();
+                timeoutTimer->deleteLater();
+
+                QDBusPendingReply<PhosphorProtocol::DragOutcome> reply = *w;
+                if (reply.isError()) {
+                    qCWarning(lcEffect) << "endDrag call failed:" << reply.error().message();
+                    return;
+                }
+                const PhosphorProtocol::DragOutcome outcome = reply.value();
+                if (const QString err = outcome.validationError(); !err.isEmpty()) {
+                    // Garbled outcome — refuse to apply any window transform.
+                    // Better to leave the window where it is than to float/snap
+                    // based on a corrupted payload.
+                    qCWarning(lcEffect) << "endDrag outcome rejected:" << err
+                                        << "— dropping without applying any action for" << windowId;
+                    return;
+                }
+                qCInfo(lcEffect) << "endDrag outcome:" << windowId << "action=" << outcome.action
+                                 << "screen=" << outcome.targetScreenId << "geo=" << outcome.toRect()
+                                 << "snapAssist=" << outcome.requestSnapAssist;
+
+                switch (outcome.action) {
+                case PhosphorProtocol::DragOutcome::NoOp:
+                case PhosphorProtocol::DragOutcome::CancelSnap:
+                case PhosphorProtocol::DragOutcome::NotifyDragOutUnsnap:
+                    // Daemon handled any internal cleanup. Nothing for the
+                    // effect to paint.
+                    break;
+
+                case PhosphorProtocol::DragOutcome::ApplyFloat: {
+                    // Autotile bypass drag ended — float the window at its
+                    // current screen. The plugin-side compositor work
+                    // (handleDragToFloat, setWindowFloatingForScreen) was
+                    // previously inlined in the dragStopped lambda; now it
+                    // fires here off the daemon's authoritative answer.
+                    //
+                    // Cross-VS transitions that happened mid-drag were
+                    // applied by slotDragPolicyChanged at the moment of
+                    // crossing, so by the time we get here the autotile
+                    // handler has the right tracking state.
+                    if (!safeWindow) {
+                        break;
+                    }
+                    const QString dropScreenId = getWindowScreenId(safeWindow);
+                    if (dropScreenId.isEmpty()) {
+                        break;
+                    }
+                    m_autotileHandler->handleDragToFloat(safeWindow, windowId, dropScreenId);
+                    // Note: m_dragFloatedWindowIds is intentionally NOT re-set here.
+                    // See dragStopped handler — the marker is cleared at drag end
+                    // because the daemon's drag-end float path (setWindowFloat →
+                    // windowFloatingStateSynced) never emits applyGeometryForFloat,
+                    // so there's nothing for the marker to suppress.
+                    PhosphorProtocol::ClientHelpers::fireAndForget(
+                        this, PhosphorProtocol::Service::Interface::WindowTracking,
+                        QStringLiteral("setWindowFloatingForScreen"), {windowId, dropScreenId, true},
+                        QStringLiteral("setWindowFloatingForScreen - endDrag ApplyFloat"));
+                    qCInfo(lcEffect) << "endDrag ApplyFloat:" << windowId << "on" << dropScreenId;
+                    break;
+                }
+
+                case PhosphorProtocol::DragOutcome::ApplySnap: {
+                    if (!safeWindow || safeWindow->isFullScreen()) {
+                        break;
+                    }
+                    const QRect snapGeometry = outcome.toRect();
+                    // If the window is still in user-move state because only
+                    // the activation mouse button is held (LMB already
+                    // released), cancel KWin's interactive move so we can
+                    // snap immediately. Without this, applySnapGeometry
+                    // defers (100ms retry) until ALL buttons are released —
+                    // noticeable delay when using a mouse button (RMB) for
+                    // zone activation.
+                    if (safeWindow->isUserMove() && !(m_currentMouseButtons & Qt::LeftButton)) {
+                        if (KWin::Window* kw = safeWindow->window()) {
+                            kw->cancelInteractiveMoveResize();
+                        }
+                    }
+                    applySnapGeometry(safeWindow, snapGeometry);
+                    break;
+                }
+
+                case PhosphorProtocol::DragOutcome::RestoreSize: {
+                    if (!safeWindow || safeWindow->isFullScreen()) {
+                        break;
+                    }
+                    // Drag-to-unsnap: apply pre-snap width/height at current
+                    // position. Skip if slotRestoreSizeDuringDrag already
+                    // applied during the drag (size within 1px).
+                    QRectF frame = safeWindow->frameGeometry();
+                    const QRect geo(static_cast<int>(frame.x()), static_cast<int>(frame.y()), outcome.width,
+                                    outcome.height);
+                    if (qAbs(frame.width() - outcome.width) <= 1 && qAbs(frame.height() - outcome.height) <= 1) {
+                        qCDebug(lcEffect) << "endDrag RestoreSize: already at correct size, skipping";
+                        break;
+                    }
+                    if (safeWindow->isUserMove() && !(m_currentMouseButtons & Qt::LeftButton)) {
+                        if (KWin::Window* kw = safeWindow->window()) {
+                            kw->cancelInteractiveMoveResize();
+                        }
+                    }
+                    // Drag-to-unsnap: window leaves zone-managed sizing, restore pre-snap dimensions.
+                    applySnapGeometry(safeWindow, geo, /*allowDuringDrag=*/false, /*skipAnimation=*/false,
+                                      PhosphorAnimation::ProfilePaths::ZoneSnapOut);
+                    break;
+                }
+                }
+
+                // Auto-fill: if window was dropped without snapping to a
+                // zone and wasn't floated, try the first empty zone on the
+                // release screen. Daemon-provided targetScreenId wins over
+                // window's current screen (cross-screen drags).
+                const bool applied = outcome.action == PhosphorProtocol::DragOutcome::ApplySnap
+                    || outcome.action == PhosphorProtocol::DragOutcome::ApplyFloat;
+                if (!applied && safeWindow && !outcome.targetScreenId.isEmpty() && isDaemonReady("auto-fill on drop")) {
+                    const bool sticky = isWindowSticky(safeWindow);
+                    auto onSnapSuccess = [this](const QString&, const QString& snappedScreenId) {
+                        m_snapAssistHandler->showContinuationIfNeeded(snappedScreenId);
+                    };
+                    tryAsyncSnapCall(PhosphorProtocol::Service::Interface::Snap, QStringLiteral("snapToEmptyZone"),
+                                     {windowId, outcome.targetScreenId, sticky}, safeWindow, windowId, true, nullptr,
+                                     onSnapSuccess);
+                }
+
+                // Snap Assist: show the window picker if the daemon
+                // requested it. asyncShow is non-blocking.
+                if (outcome.requestSnapAssist && !outcome.emptyZones.isEmpty() && !outcome.targetScreenId.isEmpty()) {
+                    m_snapAssistHandler->asyncShow(windowId, outcome.targetScreenId, outcome.emptyZones);
+                }
+            });
+}
+
+void PlasmaZonesEffect::callCancelSnap()
+{
+    qCInfo(lcEffect) << "Calling cancelSnap (drag cancelled by Escape or external event)";
+    // QDBusMessage::createMethodCall — purely local, no D-Bus introspection.
+    QDBusMessage msg =
+        QDBusMessage::createMethodCall(PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                                       PhosphorProtocol::Service::Interface::WindowDrag, QStringLiteral("cancelSnap"));
+    QDBusConnection::sessionBus().asyncCall(msg);
+}
+
+void PlasmaZonesEffect::tryAsyncSnapCall(const QString& interface, const QString& method, const QList<QVariant>& args,
+                                         QPointer<KWin::EffectWindow> window, const QString& windowId,
+                                         bool storePreSnap, std::function<void()> fallback,
+                                         std::function<void(const QString&, const QString&)> onSnapSuccess,
+                                         bool skipAnimation, std::function<void()> onComplete)
+{
+    QDBusPendingCall call = PhosphorProtocol::ClientHelpers::asyncCall(interface, method, args);
+    auto* watcher = new QDBusPendingCallWatcher(call, this);
+    connect(watcher, &QDBusPendingCallWatcher::finished, this,
+            [this, window, windowId, storePreSnap, method, fallback, onSnapSuccess, args, skipAnimation,
+             onComplete](QDBusPendingCallWatcher* w) {
+                w->deleteLater();
+                QDBusPendingReply<int, int, int, int, bool> reply = *w;
+                if (reply.isError()) {
+                    qCDebug(lcEffect) << method << "error:" << reply.error().message();
+                    if (fallback)
+                        fallback();
+                    if (onComplete)
+                        onComplete();
+                    return;
+                }
+                if (reply.argumentAt<4>() && window) {
+                    QRect geo(reply.argumentAt<0>(), reply.argumentAt<1>(), reply.argumentAt<2>(),
+                              reply.argumentAt<3>());
+                    qCInfo(lcEffect) << method << "snapping" << windowId << "to:" << geo;
+                    if (storePreSnap)
+                        ensurePreSnapGeometryStored(window, windowId, window ? window->frameGeometry() : QRectF());
+                    applySnapGeometry(window, geo, false, skipAnimation);
+                    // args[1] is screenId (e.g. for snapToEmptyZone, snapToLastZone)
+                    if (onSnapSuccess && args.size() >= 2) {
+                        onSnapSuccess(windowId, args[1].toString());
+                    }
+                    if (onComplete)
+                        onComplete();
+                    return;
+                }
+                if (fallback)
+                    fallback();
+                if (onComplete)
+                    onComplete();
+            });
+}
+
+void PlasmaZonesEffect::repaintSnapRegions(KWin::EffectWindow* window, const QRectF& oldFrame, const QRect& newGeo)
+{
+    window->addRepaintFull();
+    // Guard the global compositor repaint requests: this method can run
+    // from late D-Bus reply callbacks (callEndDrag → applySnap → here)
+    // that may dispatch during compositor teardown, when KWin::effects
+    // has been torn down. The window-local addRepaintFull above is
+    // safe because the EffectWindow itself is alive (we hold a
+    // QPointer-checked reference at the call site).
+    if (KWin::effects) {
+        if (oldFrame.isValid()) {
+            KWin::effects->addRepaint(oldFrame.toAlignedRect());
+        }
+        KWin::effects->addRepaint(newGeo);
+    }
+}
+
+void PlasmaZonesEffect::applySnapGeometry(KWin::EffectWindow* window, const QRect& geometry, bool allowDuringDrag,
+                                          bool skipAnimation, const QString& profilePath)
+{
+    if (!window) {
+        qCWarning(lcEffect) << "applyGeometry: window is null";
+        return;
+    }
+
+    // Normalize so width/height are non-negative; reject invalid rects
+    QRect geo = geometry.normalized();
+    if (!geo.isValid() || geo.width() <= 0 || geo.height() <= 0) {
+        qCWarning(lcEffect) << "applyGeometry: invalid or empty geometry:" << geometry;
+        return;
+    }
+
+    // Don't call moveResize() on fullscreen windows, it can crash KWin.
+    // See KDE bugs #429752, #301529, #489546.
+    if (window->isFullScreen()) {
+        qCDebug(lcEffect) << "applyGeometry: window is fullscreen, skipping";
+        return;
+    }
+
+    // For X11/XWayland windows, KWin constrains the frame size to align with
+    // WM_SIZE_HINTS (size increments for terminals like Ghostty, Kitty, etc.).
+    // Pre-compute the constrained size and center the window in its zone so the
+    // gap is distributed evenly instead of all at the bottom-right.
+    // This applies to all snap operations (zone snap, autotile, resnap, etc.).
+    // Wayland-native clients negotiate size async (constrainFrameSize only
+    // checks min/max, not char-cell grid), so they're handled by the deferred
+    // check in slotWindowFrameGeometryChanged().
+    if (window->isX11Client()) {
+        KWin::Window* kw = window->window();
+        if (kw) {
+            const QSizeF constrained = kw->constrainFrameSize(QSizeF(geo.size()));
+            const int cw = qRound(constrained.width());
+            const int ch = qRound(constrained.height());
+            if (cw < geo.width() || ch < geo.height()) {
+                // Clamp to non-negative: if min-size exceeds the zone in one
+                // dimension, don't shift the window beyond the zone's edge.
+                const int dx = qMax(0, geo.width() - cw) / 2;
+                const int dy = qMax(0, geo.height() - ch) / 2;
+                geo = QRect(geo.x() + dx, geo.y() + dy, cw, ch);
+                qCDebug(lcEffect) << "Pre-centered X11 window with size constraints:"
+                                  << "zone=" << geometry.size() << "constrained=" << constrained << "adjusted=" << geo;
+            }
+        }
+    }
+
+    // Skip no-op: if window is already at the target geometry AND there is
+    // no in-flight animation, calling moveResize() is redundant and can have
+    // subtle stacking side effects on some KWin versions (e.g. during daemon
+    // restart double-processing).
+    //
+    // When an animation IS in flight, frameGeometry() already reflects the
+    // committed target from the previous applySnapGeometry's moveResize —
+    // but the visual position is still mid-transition. A rapid reversal
+    // (float → unfloat, rotate → rotate back) legitimately targets the same
+    // committed geometry and must NOT be skipped, because the animation needs
+    // to play from the current visual position to that target.
+    if (QRectF(geo) == window->frameGeometry() && !m_windowAnimator->hasAnimation(window)) {
+        qCDebug(lcEffect) << "moveResize: window already at target geometry, skipping:" << geo;
+        return;
+    }
+
+    qCDebug(lcEffect) << "Setting window geometry from" << window->frameGeometry() << "to" << geo;
+
+    // Capture old frame before moveResize for repaint region
+    const QRectF oldFrame = window->frameGeometry();
+
+    // In KWin 6, we use the window's moveResize methods
+    // When allowDuringDrag is false: defer if window is in user move/resize (snap on release)
+    // When allowDuringDrag is true: apply immediately (snap-on-hover during drag)
+    if (!allowDuringDrag && (window->isUserMove() || window->isUserResize())) {
+        qCDebug(lcEffect) << "Window in user move/resize, deferring geometry via windowFinishUserMovedResized";
+        QPointer<KWin::EffectWindow> safeWindow = window;
+        auto conn = std::make_shared<QMetaObject::Connection>();
+        *conn = connect(window, &KWin::EffectWindow::windowFinishUserMovedResized, this,
+                        [this, safeWindow, geo, skipAnimation, profilePath, conn](KWin::EffectWindow*) {
+                            disconnect(*conn);
+                            if (safeWindow && !safeWindow->isDeleted() && !safeWindow->isFullScreen()) {
+                                applySnapGeometry(safeWindow, geo, false, skipAnimation, profilePath);
+                            }
+                        });
+        return;
+    }
+
+    // Animation: moveResize to the final geometry immediately, then morph
+    // the window visually from its old position/size to the new one using
+    // translate + scale in paintWindow(). This follows the standard KDE
+    // effect pattern — effects are visual overlays, never per-frame moveResize.
+    if (!skipAnimation && !allowDuringDrag && m_windowAnimator->isEnabled()) {
+        const QRectF targetFrame(geo);
+
+        // Apply final geometry immediately — client starts re-rendering at new size.
+        // Do this before touching the animator so the controller's
+        // downstream bounds / padding queries see the updated
+        // expandedGeometry for this frame.
+        KWin::Window* kw = window->window();
+        if (kw) {
+            kw->moveResize(targetFrame);
+        }
+
+        if (m_windowAnimator->hasAnimation(window)) {
+            if (m_windowAnimator->isAnimatingToTarget(window, targetFrame)) {
+                return; // Already animating to this target
+            }
+            // Capture the displaced animation's endpoints before retarget
+            // modifies or deletes the entry. On a rapid reversal where
+            // advance() hasn't ticked, m_current still equals m_from
+            // (the animation's start point), so retarget(newTarget) sees
+            // current ≈ newTarget when the reversal goes back to the
+            // original zone — degenerate. Use the displaced animation's
+            // TARGET as the visual origin for the replacement: that's
+            // where the window was visually heading (and where moveResize
+            // just committed to), so animating from there to the new
+            // target matches the user's expectation.
+            const QRectF displacedTarget = m_windowAnimator->animationFor(window)->to();
+            const QRectF visualPos = m_windowAnimator->currentValue(window, QRectF(oldFrame));
+            const auto result = m_windowAnimator->retargetWithResult(
+                window, targetFrame, PhosphorAnimation::RetargetPolicy::PreserveVelocity);
+            if (result == PhosphorAnimation::RetargetResult::DegenerateReap) {
+                // Retarget collapsed (current visual ≈ new target).
+                // Start a fresh animation from the displaced target
+                // (where the window was heading) to the new target.
+                // If that's also degenerate (same point), startAnimation
+                // returns false and no animation plays — correct, since
+                // there's no visual distance to cover.
+                const QRectF animFrom = (displacedTarget != targetFrame) ? displacedTarget : visualPos;
+                m_windowAnimator->startAnimation(window, animFrom, targetFrame);
+            }
+        } else {
+            m_windowAnimator->startAnimation(window, QRectF(oldFrame), targetFrame);
+        }
+
+        if (m_windowAnimator->hasAnimation(window)) {
+            const auto shaderProfile = m_shaderProfileTree.resolve(profilePath);
+            if (!shaderProfile.effectiveEffectId().isEmpty()) {
+                beginShaderTransition(window, shaderProfile);
+            }
+        }
+
+        repaintSnapRegions(window, oldFrame, geo);
+        return;
+    }
+
+    // No animation path (disabled, during drag, etc.): apply moveResize directly.
+    if (m_windowAnimator->hasAnimation(window)) {
+        m_windowAnimator->removeAnimation(window);
+    }
+
+    KWin::Window* kwinWindow = window->window();
+    if (kwinWindow) {
+        qCInfo(lcEffect) << "moveResize: QRect=" << geo << "-> QRectF=" << QRectF(geo);
+        kwinWindow->moveResize(QRectF(geo));
+
+        repaintSnapRegions(window, oldFrame, geo);
+    } else {
+        qCWarning(lcEffect) << "Cannot get underlying Window from EffectWindow";
+    }
+}
+
+void PlasmaZonesEffect::slotRestoreSizeDuringDrag(const QString& windowId, int width, int height)
+{
+    // Restore pre-snap size when cursor leaves zone during drag. The window may have been
+    // snapped when the drag started (at zone size); when the user drags out of all zones,
+    // we restore to floated state immediately so they see the window return to original size.
+    // This complements the release path (dragStopped) which also handles restore.
+    if (!m_dragTracker->isDragging() || m_dragTracker->draggedWindowId() != windowId) {
+        return;
+    }
+
+    KWin::EffectWindow* window = m_dragTracker->draggedWindow();
+    if (!window || !shouldHandleWindow(window)) {
+        return;
+    }
+
+    if (width <= 0 || height <= 0) {
+        return;
+    }
+
+    // Restore-size-only: keep current position, apply pre-snap width/height
+    QRectF frame = window->frameGeometry();
+    QRect geometry(static_cast<int>(frame.x()), static_cast<int>(frame.y()), width, height);
+
+    qCDebug(lcEffect) << "Restoring size during drag:" << windowId << geometry;
+    // Live drag-out unsnap: restoring pre-snap dimensions while the user is still dragging.
+    // Logically a snap-out (the window is leaving zone-managed sizing).
+    applySnapGeometry(window, geometry, /*allowDuringDrag=*/true, /*skipAnimation=*/false,
+                      PhosphorAnimation::ProfilePaths::ZoneSnapOut);
+}
+
+void PlasmaZonesEffect::slotSnapAssistReady(const QString& windowId, const QString& releaseScreenId,
+                                            const PhosphorProtocol::EmptyZoneList& emptyZones)
+{
+    // Discard if a new drag has already started — this signal was from a
+    // prior drop. The daemon defers the compute to after endDrag returns,
+    // so by the time this slot fires the user may already be dragging again.
+    if (m_dragTracker->isDragging()) {
+        qCDebug(lcEffect) << "Discarding snapAssistReady: new drag in progress";
+        return;
+    }
+    if (emptyZones.isEmpty() || releaseScreenId.isEmpty()) {
+        return;
+    }
+    m_snapAssistHandler->asyncShow(windowId, releaseScreenId, emptyZones);
+}
+
+void PlasmaZonesEffect::slotDragPolicyChanged(const QString& windowId, const PhosphorProtocol::DragPolicy& newPolicy)
+{
+    // Daemon-owned cross-VS flip. The daemon's updateDragCursor
+    // handler computed policy at the current cursor position and found it
+    // different from the policy in force — tell us so we can apply the
+    // compositor-level transition. Replaces the effect-side cross-VS flip
+    // loop in the dragMoved lambda that walked KWin::effects->screens()
+    // with a stale m_autotileScreens cache.
+    //
+    // Guards: this slot only acts if we're actively tracking the drag for
+    // this windowId. Stray signals (daemon restart, out-of-order delivery)
+    // are ignored.
+    if (!m_dragTracker->isDragging() || m_dragTracker->draggedWindowId() != windowId) {
+        qCDebug(lcEffect) << "slotDragPolicyChanged: drag no longer active for" << windowId;
+        return;
+    }
+
+    if (const QString err = newPolicy.validationError(); !err.isEmpty()) {
+        // Garbled policy change — keep current state rather than transitioning
+        // to a corrupted one. The daemon will re-emit on the next cursor tick
+        // if this was transient.
+        qCWarning(lcEffect) << "slotDragPolicyChanged rejected:" << err << "for" << windowId;
+        return;
+    }
+
+    const PhosphorProtocol::DragBypassReason oldReason = m_currentDragPolicy.bypassReason;
+    const PhosphorProtocol::DragBypassReason newReason = newPolicy.bypassReason;
+    if (oldReason == newReason) {
+        // Same reason but different screenId (autotile→autotile cross-VS):
+        // update the captured screen so endDrag's ApplyFloat uses the right one.
+        m_currentDragPolicy = newPolicy;
+        if (newReason == PhosphorProtocol::DragBypassReason::AutotileScreen) {
+            m_dragBypassScreenId = newPolicy.screenId;
+        }
+        return;
+    }
+
+    qCInfo(lcEffect) << "slotDragPolicyChanged:" << windowId << oldReason << "->" << newReason
+                     << "screen=" << newPolicy.screenId;
+
+    m_currentDragPolicy = newPolicy;
+
+    KWin::EffectWindow* dragW = m_dragTracker->draggedWindow();
+
+    if (newReason == PhosphorProtocol::DragBypassReason::AutotileScreen) {
+        // Snap → autotile (or context-disabled → autotile). Cancel any
+        // active snap overlay, enter bypass mode. Mirrors the old
+        // effect-side flip block's "snap→autotile" branch, but driven by
+        // daemon truth rather than an effect-cached screen set.
+        if (!m_dragBypassedForAutotile) {
+            callCancelSnap();
+            m_dragBypassedForAutotile = true;
+            m_dragBypassScreenId = newPolicy.screenId;
+            m_dragStartedSent = false;
+            m_pendingDragWindowId.clear();
+            m_pendingDragGeometry = QRectF();
+            m_snapDragStartScreenId.clear();
+        } else {
+            // Already in bypass but on a different autotile screen — just
+            // update the captured screen id.
+            m_dragBypassScreenId = newPolicy.screenId;
+        }
+        return;
+    }
+
+    if (oldReason == PhosphorProtocol::DragBypassReason::AutotileScreen) {
+        // Autotile → snap (or autotile → context-disabled). Drop the
+        // bypass flag and initialize snap-drag state as if the drag just
+        // started on this snap screen. Remove the window from autotile
+        // tracking so slotWindowFrameGeometryChanged doesn't fight the
+        // snap geometry on subsequent geometry changes.
+        //
+        // Do NOT call handleDragToFloat here: the mid-drag schedule would
+        // race against the zone snap at drop, making the window jump after
+        // the user lets go. onWindowClosed alone clears the tracking state.
+        if (dragW) {
+            m_autotileHandler->onWindowClosed(windowId, m_dragBypassScreenId);
+        }
+        m_dragBypassedForAutotile = false;
+        m_dragActivationDetected = false;
+        m_dragStartedSent = false;
+        m_pendingDragWindowId = windowId;
+        m_pendingDragGeometry = dragW ? dragW->frameGeometry() : QRectF();
+        m_snapDragStartScreenId = newPolicy.screenId;
+        if (!m_keyboardGrabbed) {
+            KWin::effects->grabKeyboard(this);
+            m_keyboardGrabbed = true;
+        }
+        return;
+    }
+
+    // Other transitions (snap ↔ context_disabled / snapping_disabled):
+    // no compositor-level work needed. The daemon will return a NoOp at
+    // endDrag for disabled paths.
+}
+
+} // namespace PlasmaZones
